requirements.txt
streamlit
langchain
langchain-openai
langchain-community
pydantic
python-dotenv
reportlab
PyPDF2
faiss-cpu
openai
